Made sure the update method only runs ont the server.
Not sure if it fixes it, but the client shouldn't be handling that
stuff anyway. All the other edits were the formatter doing its thing.
Let me know if this changes anything. Also, the plants were with a
weird colour because I made them pink on purpose. I thought the erebus
needed some more of girl power.