LEX! PLAY WITH THIS!
Added some basic code for you to play with.
Shift Click with the Caveman's club.....
You have a random height, radius and roots!
If you can - Add random branches and make the top few blocks of the
trunk  all broken and random and maybe a few holes here and there....