Closes #402
The portal is now made by filling a vanilla portal shape with
leaves, placing a Gaean Keystone on the floor in front, and clicking the
keystone with the Staff of Gaea. The staff is consumed instead of damaged.
Another keystone is provided on the other side, but if you wish to
easily return to the overworld without passing through the land of the
dead, you will need a second Staff and more leaves.

Traveling through the portal places you on the keystone
in the other dimension.

Old portals should be safe to use so long as they are in a reasonably
openish area; in particular there should be free space above the
keystone.

----------------
----------------

The extra fancy stuff mentioned in the issue is not implemented.
(particles, bug spawns, animated portal creation)

The portalblock rules don't make it self-collapse, it has to be done the
hard way.

I'm not super-sure about entity placement; there seems to be some magic
going on that I couldn't find the source of.

The keystone's texture/model might use some tweaking to match the
hovering staff thing.

The staff-is-consumed thing might cause too much excitement.