Improvements to the lang file.
I'll make it a pull request so you can read over it. Changed some names (for the better, in my opinion), shortened some things and cleaned it up a bit. Made some tooltips a bit more explicit. Fixed the Bog Maw having no name. Some further minor changes. Changed Bamber to Bambi by the way, not sure if that gets us in legal troubles but I think Bambi has a similar wit to it as Chester.